requirements.txt
environs==9.5.0
Django==4.1.7
django-extensions==3.2.1
phonenumbers==8.13.8
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
Pillow==9.5.0
django-filter==23.2
more-itertools==9.1.0
